jesus-backend-flask/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo 2.3.0
Flask-WTF==1.2.1
itsdangerous  2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==24.0
pymongo==4.7.3
Werkzeug==3.0.3
WTForms==3.1.2
python-dotenv==1.0.1